requirements.txt
ta
pandas
requests
python-binance
